Create common event form component to has create/edit event form fields and logic
